refactor(search-exp): core/ package + __main__-guarded entry points
- Move the six machinery modules (adaptive, broad_search, judge, ranking,
  snowball, source) into core/ so the dir root distinguishes entry points
  from library code; imports + doc path references updated (line-number
  cites unchanged — file contents untouched by the move).
- Guard every executable script with __main__ so imports are side-effect
  free (an unguarded top-level asyncio.run meant 'import run_experiment'
  started the full multi-hour run); add python -m entry points to
  reporting/{collect_results,cost,plots} so each stage is a plain uv run.
- 183 tests pass; REPL-first run_xxx() convention unchanged.
