+++ backend/testing/r_and_d/search_experiments/core/__init__.py
@@ -1,0 +1,1 @@
+"""Core experiment machinery: judging, adaptive loading, snowball, ranking, orchestration."""
