+++ backend/testing/r_and_d/search_experiments/retrieval/enrich.py
@@ -33,7 +33,7 @@ from __future__ import annotations
 import logging
 from dataclasses import dataclass
 
-from source import Candidate
+from core.source import Candidate
 
 logger = logging.getLogger(__name__)
