+++ backend/testing/r_and_d/search_experiments/retrieval/suggest.py
@@ -36,7 +36,7 @@ from pydantic import BaseModel, Field
 from _backend import get_llm
 from config import CONFIG
 from retrieval import _cache
-from source import Candidate
+from core.source import Candidate
 
 logger = logging.getLogger(__name__)
