+++ backend/testing/r_and_d/search_experiments/smoke/phase4_openalex.py
@@ -19,7 +19,7 @@ from _bootstrap import rule
 from retrieval import _cache
 from retrieval.enrich import classify_text_basis
 from retrieval.openalex_client import OpenAlexSource
-from snowball import build_edges, promote_snowball
+from core.snowball import build_edges, promote_snowball
 
 CONTENT = "effect of free school meals on educational attainment in the UK"
